separate PEER code
