+++ praxis/configuration_praxis.py
@@ -17,9 +17,9 @@ class PraxisConfig(PretrainedConfig):
         vocab_size=4096,
         context_length=4096,
         activation="mish",
-        peer_heads=2,
+        peer_heads=4,
         peer_experts=128**2,
-        peer_experts_per_head=4,
+        peer_experts_per_head=8,
         sparse=False,
         shuffle=False,
         pad_token_id=0,
